colors

+++ topo/tools.go
@@ -3,7 +3,7 @@ package topo
 import "errors"
 
 // GetAltitude get an altitude in the buffer
-func (buffer Buffer) GetAltitude(x uint32, y uint32) int16 {
+func (buffer Buffer) GetAltitude(x int, y int) int16 {
 	return buffer.Data[y*buffer.Width+x]
 }
 
@@ -12,8 +12,8 @@ func (buffer Buffer) SetScale(scale float32) (Buffer, error) {
 	if scale > 1 {
 		return Buffer{}, errors.New("scale must be lower than 1")
 	}
-	width := uint32(float32(buffer.Width) * scale)
-	height := uint32(float32(buffer.Height) * scale)
+	width := int(float32(buffer.Width) * scale)
+	height := int(float32(buffer.Height) * scale)
 	newBuffer := Buffer{
 		width,
 		height,
@@ -23,9 +23,9 @@ func (buffer Buffer) SetScale(scale float32) (Buffer, error) {
 		buffer.Diff,
 	}
 
-	for y := uint32(0); y < height; y++ {
-		for x := uint32(0); x < width; x++ {
-			newBuffer.Data[y*newBuffer.Width+x] = buffer.GetAltitude(uint32(float32(x)/scale), uint32(float32(y)/scale))
+	for y := 0; y < height; y++ {
+		for x := 0; x < width; x++ {
+			newBuffer.Data[y*newBuffer.Width+x] = buffer.GetAltitude(int(float32(x)/scale), int(float32(y)/scale))
 		}
 	}
